Added Example of the "Circle" class and its test class
The Kreis class is used to define a simple class for representing circle object's.
The test class uses this Kreis class and creates some circles before it ask the user to enter radius and radius enlargement factor via a JInputDialog. It then creates a table of 30 rows using this values and shows the radius, its extent and its area for each.